src/requirements.txt
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.1
click==7.1.1
flask==1.1.1
itsdangerous==1.1.0
py-healthcheck==1.9.0
six==1.14.0
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
gunicorn==19.10.0
redis==3.4.1